requirements/production.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.7
django-appconf

# Templates
django_compressor
markdown
docutils
typogrify

# Security
bleach
#cef

# Logging
#raven

# Celery: Message queue
# celery
# django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
# python-memcached

# Geo stuff
geopy
django-leaflet
django-geojson

# Admin
Werkzeug
ipython
django-debug-toolbar
django-grappelli
django-filebrowser

django-extensions
django-markup
django-reversion

django-braces
django-taggit
django-dajaxice

git+https://github.com/bread-and-pepper/django-userena.git@5bbf2749b8c9f3bc38206de11ed2fc507b9bb089
git+https://github.com/etianen/django-watson.git@de5fcb3064a427b6dbdf04a233d2adc6748104f9

pillow
